Move tracing rate limiters to Core
Dynamic instrumentation will use the TokenBucket
rate limiter. To facilitate this, the basic rate limiters
are being moved from tracing to core.
